requirements.txt
librosa==0.10.1
numpy==1.25.2
openai==1.3.0
opencv-python-headless
PyGetWindow==0.0.9
python-dotenv==1.0.0
Requests==2.31.0
soundfile==0.12.1
streamlit==1.28.1
packages/streamlit_webrtc_display_capture-0.47.1.tar.gz